fix: 딜러 통계 화면 개선 및 consult dealer_no 타입 오류 수정

+++ src/main/java/com/kiaev/client/consult/Consult.java
@@ -26,7 +26,7 @@ public class Consult {
     private Long carNo;
 
     @Column(name = "dealer_no")
-    private Long dealerNo;
+    private Integer dealerNo;
 
     @Column(name = "budget_amount")
     private Integer budgetAmount;
